requirements.txt
Django==1.11.5
django-crispy-forms==1.7.0
djangorestframework==3.7.1
mysqlclient==1.3.12
olefile==0.44
Pillow==4.2.1
pytz==2017.2
